Ask for overwrite output files when disassembling. Some log stuff.

+++ src/ui/workers/kworker.js
@@ -4,18 +4,19 @@ const pathIsInside = require('../path-is-inside');
 const krakatau = require('../krakatau');
 
 
-function pathIsValid(path) {
-    return /^([a-z]:\\|\/)[a-z0-9_]+[\\/a-z0-9_ .]*$/i.test(path);
-}
-
-function disassemble(inputFile, outputDir, roundtrip) {
+function disassemble(inputFile, outputDir, roundtrip, overwrite) {
     const data = fs.readFileSync(inputFile);
     const result = krakatau.disassemble(data, roundtrip);
     const clsName = result.name || ('no_class_name_' + new Date().getTime());
     let outPath = path.resolve(outputDir, ...clsName.split('/')) + '.j';
-    if (!pathIsValid(outPath) || !pathIsInside(outPath,outputDir) || outPath.length > 255 )
+    if (!pathIsInside(outPath,outputDir) || outPath.length > 255 )
         outPath = path.resolve(outputDir, path.basename(inputFile).replace(/\.class$/,'')) + '.j';
-    fs.writeFileSync(outPath, result.data, {flag: "wx"});
+    if (overwrite) fs.writeFileSync(outPath, result.data);
+    else {
+        if (fs.existsSync(outPath))
+            outPath = outPath.replace(/\.j$/, '_' + (new Date().getTime()) + '.j');
+        fs.writeFileSync(outPath, result.data, {flag: "wx"});
+    }
     return outPath;
 }
 
@@ -30,7 +31,7 @@ function assemble(inputFile, outputDir) {
             let {name, data} = result;
             if (!name || name.length === 0) name = ('no_class_name_' + new Date().getTime());
             let outPath = path.resolve(outputDir, ...name.split('/')) + '.class';
-            if (!pathIsValid(outPath) || !pathIsInside(outPath,outputDir) || outPath.length > 255 )
+            if (!pathIsInside(outPath,outputDir) || outPath.length > 255 )
                 outPath = path.resolve(outputDir, path.basename(inputFile).replace(/\.j$/,'_CLASS_' + nClass)) + '.class';
             fs.writeFileSync(outPath, data);
             eventResult.push({ success:true, name:name, outputFile:outPath });
@@ -51,7 +52,7 @@ onmessage = (event) => {
     try {
         switch (event.data.op){
             case 'disassemble':
-                postMessage({ success:true, result:disassemble(event.data.inputFile, event.data.outputDir, event.data.roundtrip) });
+                postMessage({ success:true, result:disassemble(event.data.inputFile, event.data.outputDir, event.data.roundtrip, event.data.overwrite) });
                 break;
             case 'assemble':
                 postMessage({ success:true, result:assemble(event.data.inputFile, event.data.outputDir) });
